Moved directory delete into the client
